Moved test_make_bids_compliant_locizer & test_make_spm_auditory to test_datasets
 Changed the tutorial text in example plot_single_subject_run.py
  - Removed mention of events.csv .
  - Removed mention of events == paradigms.
